Add adding activities, sets, and cycling set state
This commit organizes some interface issues concerning Firebase
Firestore and how data is modelled. The organization and layout of the
primary interfaces for this application is a work in progress. The ideal
arrangement of collections and arrays within Firestore is unclear.
This is something there may be best practices on that I could use.

Most of the work (and users time) is spent in StartTraining.tsx.

Sets (`ActivitySet`s) can be added to Activity instances via the UI now.
I am not sure where I want that `+` button placed - left or right. In
the gym I think my left hand access the phone more often, this may be
due to right-handedness. The button could technically be on both sides.
Maybe double-clicking the activity name should add a set for that item.

There is general design work to be done.

A fake date is used to display the datetime in the top-left corner when
writing logs.

This commit disables TSconfig's `noImplicitAny` rule. This was disabled
because it would cause verbose code when destructuring many properties
on a `.map` inside a render function.