Add Account page with clickable log history
This is the big cheese.
You can now view all logs in an account (review your training history).
Each training log can be clicked on (listed by timestamp) to view the
Start Training editor so you can get after it.

`date-fns` was added to `format` timestamps for display.

I (quickly) decided that logs viewed from log history cannot be deleted
from a user's log history in the Training editor. Users may want this
in the future.

This commit update `useDateState` to allow its data-fetching function
to return `DataState<T>` instead of `T` due to the one of the unions
being just `T` itself. This allows for a variant of DataState to be
returned instead of some instance of `T` or being forced to add `null`.