Add ability to rename sets
This commit adds set renaming and adds the ability to delete existing
training logs. Previously only training logs that have not been saved
could be deleted.

The UI layout is not final, elements will continue to move around to
find the ideal interface.

This commit also adds the index to each set display.