Add duplicating sets + Can scroll Account page
This commit makes the list of training logs on the Account page
scrollable. This can continue to be improved in the future as the
header of the Account page should not scroll with the list of logs.

Duplicating sets has been added. I think this is a feature that will be
useful to me in the case where I do not want to re-type a set's name but
I want to do that set again.

There was a redundant set of `<Columns />` in Account page, these have
been flattened.

Date formats may be made into enum variants in the future for
consistency's sake.

When prompting the user (to update the set or activity name), default
values have been removed. AKA no longer will the user see the already
existing set/activity name in the input prompt.

The "delete" option for sets has been emphasized in bold to portray the
seriousness of the action.