Add training log titles
The changes in this commit allow authenticated users to input a summary
title when starting training. During training the title is visible at
the top center of the device screen. It can be updated by clicking it
and inputting a non-empty title. The input is not vetted or tested for
appropriateness of any kind.

The NPM package `@emotion/css` has been added. Inline styles have been
desired in this project (and others) for some time. I am happy that
emotion offers @emotion/_styled_ as a stand-in for the styled-components
API. The switch was a matter of search-and-replace. (Which reminds me I
need to stop using VSCode.)

There was a typo in the `ActivityStatus.Skipped` enum variant that has
been fixed.

It is unclear how the Account page should display logs for the user. But
the current solution is absolutely not it :).

A `!!user` condition has been added in the fetch function for `logDate`
in hopes that that will solve the error that's been coming up.

Several functions, such as `exitTraining` have and will be pulled out of
their in-line definitions. This is an effort to increase code clarity
for future readers, myself included.

A flimsy console.log made it into the reposity, a rule should be added
to never allow it again. (Linter?)