Update Account UI and fix renaming log titles
This commit freshens up the logs view on the account page. Both the
title and log timestamp should be more readable. Logs are important
entities for users, they should have a sort-of "big" UI to each of them
and be card-like.

Some loading|error states have been changed for DataStateView usage
sites. There may be a useful default error and loading UI to put in
place for places where it is not important.

For now, empty titles are given to logs. There is no indicator that log
titles can be added without cluttering the UI with some default title.

There is a big movement in code coming soon. The "Go" page must be split
out from the actual Training UI. This will reduce complexity.

Additionally, remove the title input from the "Go" page.