Add log replayability on Account view
It works! The convoluted concept of having a separate collection for
logs being used as templates is too much overhead. Implementing features
the simplest way possible first reduces a lot of wasted time. It was
unclear how this feature would work - it could be done in a handful of
ways. This was easier to see how the feature feels first before worrying
about coding this and that. Less code is better!