Add weight to sets
Instead of passing `logId` as a prop through components that need it,
now that the log ID is guaranteed to be present in the url, we can use
the `useParams` hook everywhere `logId` is needed.

Also, removed `repCount` field from `ActivitySet`. This has not been
touched yet and will handled/implemented at a later time.