Add ability to move activity order up and down
This is an exciting feature.

The implementation is messy. The basic idea is get references to both
activity documents, create a batch (so we are not charged for multiple
writes in Firebase), and update both with new `position` values.

One thing that can be improved in the future (soon) is in `addActivity`.
Currently, every time an activity is added, a call to fetch the previous
activity document is made in order to calculate the new activity's
`position`. This can be switched to use the `activities` array if
`ActivitiesListView` and `StartTraining` FCs are combined.

I am wondering if all of the callbacks in `ActivityView` should be made
into `useCallback` equivalents for rendering/optimization.