Notes can be multiline & empty notes are hidden
This commit switches notes input from `input` to `textarea`. `textarea`
unfortunately is not automatically resizing as input grows. This could
be implemented differently in the future but works for now.

A feature I want for notes is to hide them if there are no notes. Not
every activity will have notes. But when there are I want them displayed
in an organized out-of-the-way fashion. And I want activities without
notes not to bug me about adding notes. The main training UI should
represent a physical book as much as possible - at least for now.

In the future it may be simpler to use only a ref for the textarea
instead of a ref and a piece of state.