Separate app routes
The top-level `App` now produces a `DataStateView` of the authenticated
(or not) firebase user. The pre-existing data model for handling whether
a user is logged-in created issues due to being asynchronous but looking
like a simple property access on the firebase.auth() object.

The SignUp and LogIn pages got some long-standing questions settled --
that being, "do I need this setUser call here?" The answer is no. The
top-level auth listener in `App` subscribes to auth changes and
re-renders routes according to the new DataStateView.

There is de-duplication work to be done with the existing `useUser`. I
kind of want to poke around with the React Context API and set up a
Provider for the firebase user object for authenticated routes.