WIP Un/following users
Work in progress - social media feature. Being able to add users and see
their training logs in one's timeline.

The `/user/:userId` route has been moved to being a route which requries
an authenticated (signed-in) user to exist. This is because when viewing
a user, to be able to follow them you must be logged in. There is also
the case, now that I am thinking of it, where you might want to be able
to view a user's account without having an account. This might occur if
you can share a user account URL but do not have an account with which
to view that route/page.

The next big thing is going to be separating TrainingLogEditor into
components and making them reusable. Everytime the UI for logs is
updated or accomodates new features, the readonly view for Timeline
should reflect the same. This could be difficult, but definitely fun.