WIP Can view followed users logs on timeline page
Split ActivitiesListView into TrainingLogEditorView (the editable view)
and TrainingLogView (the readonly view). These are both thin wrappers
around ActivityView which fetch activities and display them. Interacting
and editing a TrainingLog is only possible in the TrainingLogEditorView.

There is an issue coming up where TrainingLog objects need to have the
authorId present on them. They do not currently. An update can be done
where we read every user and programmatically write the authorId onto
the object.

Styling of the TrainingLogView is yet to be done. It looks very ugly
right now. And the weight entry in its disabled state is hideous.

We are betting on native `disabled` prop on Buttons and everything else
editable in a TrainingLog to make the read-only view successful. It
would be traumatic if other users were able to edit your logs just
because they follow you.

Some comments have been added to explain parts of the codebase.