Display log titles and timestamps on Timeline
Styling (general padding) has been improved on the TrainingLogView and
as a result the Timeline page looks much more pleasing (read: less
ugly).

TrainingLogView now takes a `log` (the kind with an authorId!). This is
better than taking piecemeal pieces of a log, now we can display the log
title and timestamp. Speaking of the log timestamps, that has finally
been "normalized" via `TrainingLog.getDate`. Making a namespaced
function for this utility could be the wrong move but I like the
DRYness.

Default log titles are now "Untitled" which is very very creative.

In the future, the entire `User` object needs to be added to a
`TrainingLog` in order for TrainingLogView to display the username of
the log author.

We now call `.unwrap` on the DataState for log title on
TrainingLogEditor. This is likely unnecessary and could cause a crash?

Modify `DataState.map` such that the operating function gets to return a
DataState which can be flattened out (not nested). `unwrapOr` and
`unwrap` have been added as they've been needed before this. An overhaul
of the `DataState` system is needed for increased clarity and joy of
use.