Make TrainingLogView style consistent
Also, avoid flashing the authenticated user display name when viewing
another user account.

Also, fix width bug in editable ActivityViews; introduced by previous
commit 2780dfb.

There was an unnecessary dependency `selectedUser` in a callback.

The "Following" button is now rendered as a result of a DataStateView
instead of straight up conditions.

ImagePreview is not working the way it should be. We will switch to an
image icon instead of viewing the actual image.