Fix broken layout due to webkit support
Safari/iOS/webkit does not allow min-content/fit-content. That's
annoying. That was not even the problem. The problem was that setting
width and height of `0px` on an input (for adding media) still behaves
as if the input element had a default width of not zero. This caused the
activity name and add activity set button to be smushed together on the
left side of the page. As if there were an input elemtent rendered on
the same row. That is fixed. "View source" on the twitter home page
(after logging in) shows how to do this for all browsers, including
sucky webkit.

`useNewTraining` has been created out of the `NewTraining` page in order
for new users with no logs on their account page to begin training ASAP.

Some pre-necessary comments for coding "feeling" for a training session
have been added. Ideas are flowing!