Update UI to square-based design.
The new UI is based of a square abbreviating the name of the activity.
"Pull ups" will display as "PU" and "KB OHP" will be "KBOHP". The design
of the previous ActivityView was boring to look at and unengaging. I
always wanted tally marks like I enjoy using in physical training log
books. The tallys should be there and it is cool to see it in action,
live.

Set names have been eliminated. There is more work to do in this area.
Set names functioned previously as set notes. Notes will be introduced
in the near future.

The feature of long-pressing the "add set" button to add a completed set
has been removed. It simply was more troublesome to implement than it
was worth.

I hope the event target checking logic for the activity square div works
as intended. Clicking the activity menu icon "more-vert" should
pass-through to that event listener.

The logic for attaching attachments to activites has been removed.
Images can be uploaded but will not display in the UI.

The tally marks solution is brought to you by
https://codepen.io/engza/pen/NRNbGq.

Theme-izing the color for the app needs to happen soon. Color variables
like palevioletred should not be flying around the app inlined!

Additionally, the `noUnusedParameters` rule has been moved from Error to
Warning. This is because developing with unused variables as
compile-stopping notifications is not something I care about!