Can comments on activities
Fix DataState alternate rendering for loading and error states due to
`?.()` usage falling through when the function call returns
null/undefined.

Pull out `AppLink` into its own component. Developing an app will need
custom-but-regular AKA themed links for routing to URLs within the app.
This lets users know they're going somewhere safe to click and is styled
consistently.

Comments have been added to activites. There was a lot to decide here.
Where is the comment UI placed on an activity? Click to unviel the
comments and input or something else? Animations? Highlight the submit
button? What word to use for the placeholder and submit button? What
styling to use on the input element to fit in with the rest of the
existing app inputs and look good? There was also the question of
duplicating stateful logic that is used with the activity notes input.

Only a small subset of existing comments are to be seen during the
activity comments preview. Only three (3) can be viewed, this could be
moved up to 5 or something if desired by users. Not sure what I want
here and will need time interacting with the app to refine this usage
and appearance.

Styling `<p />` usage is a recurring problem. Font size? Font color?
These things need to be solved. Theming for the app, in general,
continues to be a place where the app could really shine that continues
to be an issue while developing other features. Having pristine colors
to work with would be as nice as it has been to use Pad all around the
application.