Add charts to activities & lock portrait mode
Charts are now shown on activities with greater than two sets and where
the second set has a non-zero weight. This makes me think about adding
rep/time count to sets. The chart could be horizontal, to the right of
the sets. That could be silly though. The stroke color of the chart is
to be decided. Colors need to be pulled out from the app into an enum
for QoL.

Moved some comments around, thought it fits better.

Moved the menu icon for the activity square further into the top-left
corner. This is so that the user's finger does not accidentally touch
the padding on the icon and fail to add a set as intended.

Also, device orientation has (hopefully) been locked to portrait mode.