Add repetition count to sets
Decided to just go for it. It has been too long since activity sets were
thought of / requested. Now we can get a taste of what it is like to
have an input to add reps shoved in the users face. This is the
beginning of being able to support timed activities. Farmers walks and
casual runs. To be determined how that will work (will graphs need a
unit type?).

Refactored `updateSetWeight` to support updating sets values in general.

And add ellipsis to the timeline search input placeholder.

Also, rename `*Css` variables to `*Style` to be consistent.

Also fixed the bug where deleted activities caused new activities to get
the wrong `position` value.