Add Past 7 Days summary on Account page
Added time constant to constants.tsx for millisecond usage.

`modulo` utility function could go elswhere, perhaps in a `library.tsx`
file. I can't believe `%` is NOT modulo operator in JavaScript!

There is an issue with sending too many fetch requests to the server
upon viewing the Account page. Separate fetches are made for the logs,
past30DayCount, and past7Days data. This can be fixed in the future by
deciding what is really necessary for the Account page to display
immediately.