Remove italic for set weight value display
Simplify font-weight usage. Had a single usage of a 500 value flying
around. Moved to 600.

Got rid of superfluous comments. Since we don't have comments
_everywhere_, they should exist where needed, not where it is not
helpful.

Added loading spinner to the activities list for the editor view. This
must have been removed because showing a loading graphic and then "you
have no items!" must not have felt smooth. But it makes sense to have it
here so I'm adding it back. I want to know why the UI area is empty, it
should tell me if something is loading!

Used less words for the "no activities" message.