Redesign user interface for activity sets
This is a big deal since the main point of interaction with the app is
adding and editing set data.

Removed activity notes. Comments supersede this functionality.

Made the color of the set data dependent on the value. If it's "0x0"
then it is gray, otherwise highlighted.

TODO has been added for the selected menu feature. This is not added in
this commit.