WIP Templates
The next step is to simplify the logDoc interface in
TrainingLogEditor.tsx. This will make it simpler to interact with
`TrainingLog | TrainingTemplate` data when rendering the editor, then we
can see the activities in the template! Success incoming!

Made `TrainingLogMenuButton` re-usable across the app. It's now used in
Account page and Editor page. This eliminates the choice I had between
"where do I put actions/controls for the logs?" because now, they're
everywhere :)

Created a TODO for myself to update the way the `db` is interacted with.
Could be simpler. Could eliminate all these `.collection(x)` calls. Made
a note for that on the dev board project in github.

Creating converters is annoying. Could automate this if it doesn't incur
cost on the client (due to invoking a generic function to create the converters programatically).

`lastUsed` can be computed on a TrainingTemplate if we grab the last log
in `logIds` (I think)