Raise NavBar icons, new templates route to Account
Also, remove a wild `styled` usage in NavBar.tsx.

I'm using the "Safari website but it's actually a web app" feature on
mobile iOS. It does not refresh the site and pull updates like a normal
website does when visited. It's a cached version of the site. This makes
updating weird...