Can view others' log templates
In the near future, users will be able to steal templates made by other
users! This is due to work in the TrainingLogMenuButton adding a "Add to
Templates" action item. I'll be able to create warm-up|exercise|stretch
routines and share it with friends and family.

Extracted TrainingLogDateView into it's own thing. This was a long time
coming due to handling different date displays in
Account/Editor/LogView pages. Happy to do the re-using thing.

Eliminated the option to start training from zero training logs in the
Account page. Either start training on the new training page or... yeah,
just do that.

Enhanced the width of the "ADD" button for activities in the Editor.