Move ActivityView UI away from square-based design
This is exciting! A new UI to interact with really changes the feel of
using the application. The ActivitySet boxes will be fun to play around
with as time goes on.

Can cycle activity set weight units between lbs and kgs.
Same thing with set rep count units, between reps and time units.

There is some de-duplication to be done in ActivityView with the
callbacks.

It's unfortunate that Firebase/Firestore does not include a way to do
migrations. I needed to add a `repCountUnit` and `weightUnit` field
(with defaulted values) onto every log. Can't do that though! Gotta
append fields in the damn editor... Looking at old logs is going to be
broken for a while ;)

Added commented-out code for set symmetry support. Not all sets are done
with full body/both hands/both feet. Sometimes it's just left side or
just right side. Need a way to denote that.