Move activity comment button to activity menu
This makes it less (easily) accessible to comment on one activity
specifically.

Commenting is meant to be social, but some training logs I've seen (in
the FIELD) do not have any activities on them... Maybe it is time to
rethink comments/socializing training. This can be solved as the app is
used more often and by more people. We'll see how this new UI feels for
now. Simpler.