Can copy template from another's templates
This feature was removed before when the TrainingLogMenuButton (and
Account page) was refactored. It also didn't work correctly before as we
had a very wrong implementation.

The core logic for copying activites has been moved out into a separate
function. This is what drives the template-copying feature. Awesome! So
glad this works! Now I can create training templates for my mom and send
them to her, less thinking for clients trying to get going!!