Use "Training Log N" for default log title
This decision was bothering me since TrainingFor existed, what should
the default title be? In my physical notebooks I do NOT care what the
title is. The content of a log will be visible on the Account page (soon
(tm)) so the title of a log does not need to communicate too much. The
title of a log ultimately does not matter except for the user to
identify it. There will be other markers in the future, like
favorites/likes on a log, the activities themselves, etc.