Create LibraryActivityView component
The Library does not care about displaying:
- Activity Tallymarks
- Activity set data (N x N)

Should the Library provide a way to bucket activites by name?
Need to implement this. Lots of little details with how the Lib will
interact with newly created activities in new logs. The data
visualization and autocomplete in the Editor are huge benefits of the
Library. Plus a more detailed account view.