Remove ability to edit/add History
This approach does not work because we are fully copying activities and
their ID's. This means many many activites (like 10+ deadlift sessions)
have the same ID's. This means they get counted as duplicates thus not
added to the history.

The approach going forward will be only new activities added to a log
will have their history added/recorded.

One challenge that needs to be solved is, when adding activities to a
log from a template, how do they get added to the history?