Copying activities from a Template get custom IDs
So, previously, when an Activity is copied to a TrainingLog from a
TrainingTemplate, the ID of that Activity would be copied as well. Or
rather, the data for the Activity would be assigned to the object under
that same ID field. So the activities were copied with the exact same
ID. That was silly.

Now, copying activities to a template or log from a template have their
own ID.

If this was implemented when templates were added, we would have been
able to implement a way to edit activity history in the Library. Live
and learn.