Give activities from template all new IDs
This enables the History/Library feature development to continue.
The issue was that multiple activities (Deadlift, for example) had the
same exact ID as the ID of the deadlift taken from the template.
Now, every activity in every log has a unique ID, unlike before.

This commit contains the commented out code that did the re-ID'ing.