Show library activities when adding new Activity
This is where all parts come together. The recent work-around hack for
fixing the activity.history interface allows activities to be fetched
from ids from activity.history. This allows activities from the library
to be loaded upon the user typing into the Add Activity input in the
Editor. Next up is to make the information displayed along with the name
of the autocompletion suggestion more useful.

This will be so cool to use in person!