Can add Activity to Log from Library
This is HUGE! :)

I'm super happy to get this done. It's satisfying to click the
autocompletion and *feel* and see the feature work for the first time.
It's smooth. Like butter. There's more to do to make it even more
buttery. But that's for later.

Moved around the `TrainingLogEditorView` to get `activities` in the
`addActivityFromLibrary` callback scope easily (lazily). I'm glad code
just works when you can move lines around like blocks.

More to do!