Use floating button for +Activity in Editor UI
This commit re-arranges the controls in the TrainingLogEditor UI.

Previously, the title sat on top of all the other controls. Below it was
the log date, add activity button, and then sleep hours.

Now, from left to right: title, date, sleep hours, add activity button.

The background color will likely be changed in the future for visual
clarity.