Creating Template from Log adds log.id to template
This feature comes from creating a template from a log, and not deleting
that log, and wanting that template to have the log (that still exists)
ID added to the template.logIds array.

This allows templates to track which logs are "part of them" when the
template is created from a log.