Allow longer text messages in toast notifications
This is in response to receiving an error message (when attempting to
add activities from the library that haven't been migrated to the new
interface spec) with extra long text. Previously, the message would be
partially cut off due to words not wrapping within the allowed width.
Now, all text can be seen, no matter the length of the message.

Also, decreased the time messages are displayed from 4 seconds to 3.