Move AddActivity btn to left of Editor panel
Also, ensure that logs without a valid `log.sleepHours` field do not
render a floating "h" due to not having a `sleepHours` field.

Moved the not-yet-clickable log date to the right side of the Editor
(bottom) panel because the left hand and left thumb is mostly doing the
clicking on a one-handed phone pickup.