Add Add Activity to Library button in Editor
Previously, when no autocomplete results were shown in the Editor when adding
an Activity, a "no results found" message would display in the autocompletion
panel.

Now, a button is presented allowing the user to create an activity with the
name they entered in the input to the saved activities library. This makes it
easier to add that activity to logs in the future, with weight unit and rep
count unit pre-filled.