Add borders to Activity input and autocomplete menu
Previously, the add activity input field and library menu had no specific
coloration to draw attention to them. Since the browser focus has to be on the
input for it to be open, it should visually pop from the remaining app UI.

Now, a Material UI input is used as the Activity input field, providing a
better placeholder/label display for the input and a nice blue bold outline.
The saved activities completion container is also bordered by a blue outline.
This helps present the information better discered from the rest of the
grey/white background coloring.