Use Chip-based view of Saved Activities in Editor
Previously, activities in the users library were rendered in the Activity
autocompletion menu as individual rows. This caused only a limited number of
items to be shown at once. Switching to a <Chip /> based view renders more
at once due to the smaller display, which is now in rows of activities.

Previously, only 4 or 5 entries could display.

Now, we are looking at around 10.

Also, fix capitalization of Activity in Add Activity input field label.