Revamp Editor control bar to dark mode/style
This commit updates the NavBar to dark background. And the Editor
bottom-aligned controls to a dark transparent gradient. Very happy with how
this turned out with the text-shadow look.

Increase NavBar height from 65 to 70px. This is in comparison to other app nav
bars (like Brave, Instagram). The nav bar icons could be larger too. Readjusted
the colors for navbar icons to fit with the darker theme too.

Made a note to get rid of TrainingLogDateView as it is only used in one place
now, and was wonky from the beginning.

Added comments to Editor render to provide some direction when swimming through
a pool of JSX.