Remove Editor footer bar
This component/UI was clunky. This commit introduces a more flexible log editor
controls UI. The add activity button sits below the list of existing activites.
That button could just as easily float above the list. The actions inside the
log title menu button can be separated out since the log title takes up the
whole top "header" of the page. The app nav bar may eventually be shortened to
just a menu icon floating top-right since the Timeline route is soon-to-be
decomissioned.